refactor(cache): consolidate cache exports and update import paths
- Export deterministicStringify from use-cache-transform main entry point
- Update cache-wrapper.ts to import from consolidated use-cache-transform export
- Simplify import paths by removing /runtime/cache-wrapper subpath references
- Update deterministic-stringify test to import from cache-wrapper re-export
- Add use-cache-transform to knip.jsonc configuration for unused dependency detection
- Remove obsolete @typescript/native-preview from ignored dependencies
- Add @napi-rs/cli to ignored dependencies for pnpm exec prepare-binaries tool
