+++ packages/use-cache-transform/src/index.ts
@@ -4,5 +4,7 @@ export type { NativeAddon, TransformOptions, TransformResult } from './native'
 export { $$cache__, encodeBoundArgs } from './runtime/cache-wrapper'
 export type { CacheEntry } from './runtime/cache-wrapper'
 
+export { deterministicStringify } from './runtime/deterministic-stringify'
+
 export { transformUseCacheModule } from './transform'
 export type { UseCacheTransformOptions } from './transform'
